Remove dep of @alephium/shared in @alephium/shared-crypto
