+++ apps/mobile-wallet/src/types/addresses.ts
@@ -16,7 +16,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { AddressHash, AddressIndex, AddressKeyPair, AddressSettings, BalanceHistory } from '@alephium/shared'
+import { AddressHash, AddressIndex, AddressSettings, BalanceHistory } from '@alephium/shared'
+import { AddressKeyPair } from '@alephium/shared-crypto'
 import { explorer } from '@alephium/web3'
 import { AddressTokenBalance } from '@alephium/web3/dist/src/api/api-explorer'
 import { EntityState } from '@reduxjs/toolkit'
