+++ apps/mobile-wallet/src/types/wallet.ts
@@ -16,7 +16,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { AddressKeyPair, AddressMetadata, Contact, ContactFormData } from '@alephium/shared'
+import { AddressMetadata, Contact, ContactFormData } from '@alephium/shared'
+import { AddressKeyPair } from '@alephium/shared-crypto'
 
 import { AddressPartial } from '~/types/addresses'
