+++ packages/shared-crypto/package.json
@@ -16,7 +16,6 @@
     "clean:node-modules": "rm -rf node_modules"
   },
   "dependencies": {
-    "@alephium/shared": "workspace:*",
     "@alephium/web3": "0.29.1",
     "base-x": "^4.0.0",
     "bcfg": "~0.1.6",
